refactor: move `LOG` into own file
preparation for future work relying on initialization order
